TSK-1925: clean up claimTaskAccTest
